pyproject.toml
[tool.poetry]
name = "nasa-cli"
version = "0.5.0"
description = "Command Line Interface made with Typer and Python for NASA API."
authors = ["Guillermo Valdes <guillermo.valdes@pjecz.gob.mx>"]
license = "AGPL3"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
typer = "^0.6.1"
rich = "^12.5.1"
python-dotenv = "^0.21.0"
Unidecode = "^1.3.5"

[tool.poetry.dev-dependencies]
pylint = "^2.15.3"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 256
